move download back to scoring script
