+++ tinyhackathon/llm_scoring.py
@@ -3,7 +3,7 @@ import os
 import re
 import traceback
 from pathlib import Path
-from typing import Annotated, Dict, List, Optional, Tuple
+from typing import Annotated, Dict, List, Optional, Tuple, Any, Union
 
 import pandas as pd
 import typer
@@ -12,11 +12,14 @@ from exllamav2.generator import ExLlamaV2DynamicGenerator, ExLlamaV2Sampler
 from rich.console import Console
 from rich.table import Table
 
+from hf_upload import get_hf_user
+
 app = typer.Typer(context_settings={"help_option_names": ["-h", "--help"]}, pretty_exceptions_show_locals=False)
 console = Console()
 
 
-def load_submissions(submissions_dir):
+def load_submissions(submissions_dir: Union[str, Path]) -> List[Path]:
+    "Load submission files from the specified directory."
     submissions_dir = Path(submissions_dir)
     if not submissions_dir.exists():
         raise ValueError(f"Submissions directory {submissions_dir} does not exist")
@@ -25,7 +28,8 @@ def load_submissions(submissions_dir):
     return files
 
 
-def process_submission(pq_file, generator, scores, output_file):
+def process_submission(pq_file: Path, generator: ExLlamaV2DynamicGenerator, scores: Dict[str, Any], output_file: Path):
+    "Process a single submission file and evaluate its completions."
     file_path = str(pq_file)
     username = pq_file.parent.name
     submission_id = pq_file.stem
@@ -46,8 +50,8 @@ def process_submission(pq_file, generator, scores, output_file):
     return scores
 
 
-def evaluate_submissions(model_dir, output_file="scores.json", submissions_dir="downloaded_submissions"):
-    "Evaluate submissions using ExLlama2"
+def evaluate_submissions(model_dir: str, output_file: str = "scores.json", submissions_dir: str = "downloaded_submissions"):
+    "Evaluate all submissions using ExLlama2."
     parquet_files = load_submissions(submissions_dir)
     console.print(f"[yellow]Loading model from {model_dir}...[/yellow]")
     config = ExLlamaV2Config(model_dir)
@@ -57,7 +61,7 @@ def evaluate_submissions(model_dir, output_file="scores.json", submissions_dir="
     tokenizer = ExLlamaV2Tokenizer(config)
     generator = ExLlamaV2DynamicGenerator(model=model, cache=cache, tokenizer=tokenizer)
     output_file = Path(output_file)
-    scores = {}
+    scores: Dict[str, Dict[str, Dict[str, Any]]] = {}
     if output_file.exists():
         scores = json.loads(output_file.read_text())
 
@@ -72,8 +76,15 @@ def evaluate_submissions(model_dir, output_file="scores.json", submissions_dir="
     return scores, leaderboard
 
 
-# Evaluate each completion
-def eval_completions(completions, generator, username, submission_id, scores, output_file):
+def eval_completions(
+    completions: List[str],
+    generator: ExLlamaV2DynamicGenerator,
+    username: str,
+    submission_id: str,
+    scores: Dict[str, Dict[str, Dict[str, Any]]],
+    output_file: Path,
+):
+    "Evaluate each completion in a submission."
     for i, completion in enumerate(completions):
         prompt = f"Rate the quality of this story completion on a scale of 1-10: {completion}"
         response = generator.generate(prompt, temperature=0.1, top_p=0.9, max_new_tokens=20)
@@ -100,7 +111,7 @@ def evaluate(
     output_file: Annotated[str, typer.Option(help="Path to save scores JSON")] = "scores.json",
     submissions_dir: Annotated[str, typer.Option(help="Directory containing submission files")] = "downloaded_submissions",
 ):
-    "Evaluate submissions using ExLlama2"
+    "Evaluate submissions using ExLlama2 and display a leaderboard."
     try:
         scores, leaderboard = evaluate_submissions(model_dir, output_file, submissions_dir)
 
@@ -121,6 +132,96 @@ def evaluate(
     except Exception as e:
         console.print(f"[red]Error: {str(e)}[/red]")
         console.print(traceback.format_exc())
+
+
+def download_new_submissions(
+    dataset_id: str = "cluster-of-stars/TinyStoriesHackathon", output_dir: Union[str, Path] = "downloaded_submissions"
+) -> List[Dict[str, Any]]:
+    "Download all new submissions from HF dataset."
+    # Get HF API
+    _, api = get_hf_user()
+
+    # Create output directory
+    output_dir = Path(output_dir)
+    output_dir.mkdir(exist_ok=True, parents=True)
+
+    # Create tracking file for processed submissions
+    processed_file = output_dir / "processed.json"
+    if processed_file.exists():
+        processed = set(json.loads(processed_file.read_text()))
+    else:
+        processed = set()
+
+    # Get all files in the dataset that match our pattern
+    files = api.list_repo_files(repo_id=dataset_id, repo_type="dataset")
+    submission_files = [f for f in files if f.startswith("submissions/") and f.endswith(".parquet")]
+
+    # Download new submissions
+    new_files = []
+    for remote_path in submission_files:
+        if remote_path in processed:
+            continue
+
+        # Extract username and filename
+        parts = remote_path.split("/")
+        if len(parts) != 3:
+            continue  # Skip unexpected formats
+
+        username = parts[1]
+        filename = parts[2]
+
+        # Create user directory
+        user_dir = output_dir / username
+        user_dir.mkdir(exist_ok=True)
+
+        # Download file
+        local_path = user_dir / filename
+        console.print(f"Downloading [yellow]{remote_path}[/yellow] to [blue]{local_path}[/blue]")
+
+        # Download to the correct path
+        downloaded_path = api.hf_hub_download(repo_id=dataset_id, repo_type="dataset", filename=remote_path, local_dir=output_dir)
+
+        # Move file to the correct location if needed
+        if Path(downloaded_path) != local_path:
+            Path(downloaded_path).rename(local_path)
+
+        # Add to processed list
+        processed.add(remote_path)
+        new_files.append(dict(username=username, filename=filename, remote_path=remote_path, local_path=local_path))
+
+    # Update processed file
+    processed_file.write_text(json.dumps(list(processed)))
+
+    return new_files
+
+
+@app.command()
+def download_submissions(
+    output_dir: Annotated[str, typer.Option(help="Directory to store downloaded submissions")] = "downloaded_submissions",
+):
+    "Download all new submissions from the TinyStories hackathon."
+    try:
+        console.print("[yellow]Downloading new submissions...[/yellow]")
+        new_files = download_new_submissions(output_dir=output_dir)
+
+        if not new_files:
+            console.print("[green]No new submissions found.[/green]")
+            return
+
+        console.print(f"[green]Downloaded {len(new_files)} new submissions:[/green]")
+
+        table = Table(show_header=True)
+        table.add_column("Username")
+        table.add_column("Filename")
+        table.add_column("Local Path")
+
+        for file in new_files:
+            table.add_row(file["username"], file["filename"], str(file["local_path"]))
+
+        console.print(table)
+
+    except Exception as e:
+        console.print(f"[red]Error: {str(e)}[/red]")
 
 
 if __name__ == "__main__":
